Decouple infra-provider from core (#78)
* Decouple infra-provider from core
* Update CHANGELOG.md
